ai_service/requirements.txt
google-generativeai
fpdf
python-dotenv